+++ src/backend_rapier3d.rs
@@ -28,6 +28,14 @@ impl Plugin for TnuaRapier3dPlugin {
     }
 }
 
+/// `bevy_rapier_3d`-specific components required for Tnua to work.
+#[derive(Bundle, Default)]
+pub struct TnuaRapier3dIOBundle {
+    pub velocity: Velocity,
+    pub external_force: ExternalForce,
+    pub read_mass_properties: ReadMassProperties,
+}
+
 /// Add this component to make [`TnuaProximitySensor`] cast a shape instead of a ray.
 #[derive(Component)]
 pub struct TnuaRapier3dSensorShape(pub Collider);
@@ -200,12 +208,26 @@ fn update_proximity_sensors_system(
     );
 }
 
-fn apply_motors_system(mut query: Query<(&TnuaMotor, &mut Velocity)>) {
-    for (motor, mut velocity) in query.iter_mut() {
-        if !motor.desired_acceleration.is_finite() {
-            continue;
+fn apply_motors_system(
+    mut query: Query<(
+        &TnuaMotor,
+        &mut Velocity,
+        &ReadMassProperties,
+        &mut ExternalForce,
+    )>,
+) {
+    for (motor, mut velocity, mass_properties, mut external_force) in query.iter_mut() {
+        if motor.lin.boost.is_finite() {
+            velocity.linvel += motor.lin.boost;
         }
-        velocity.linvel += motor.desired_acceleration;
-        velocity.angvel += motor.desired_angacl;
+        if motor.lin.acceleration.is_finite() {
+            external_force.force = motor.lin.acceleration * mass_properties.0.mass;
+        }
+        if motor.ang.boost.is_finite() {
+            velocity.angvel += motor.ang.boost;
+        }
+        if motor.ang.acceleration.is_finite() {
+            external_force.torque = motor.ang.acceleration * mass_properties.0.principal_inertia;
+        }
     }
 }